PROJECT_WE _WANT_TO_MERGE/tata_captcha-test/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
onnxruntime==1.24.4
numpy==2.4.3
Pillow==12.1.1
requests==2.32.5
jinja2==3.1.4
aiofiles==24.1.0
python-multipart==0.0.9
